BookLibrary.Application/Books/Commands/DeleteBook/DeleteBookCommandHandler.cs
using MediatR;
using BookLibrary.Domain.Entities;
using BookLibrary.Domain.Interface;

namespace BookLibrary.Application.Books.Commands.DeleteBook
{
    public class DeleteBookCommandHandler : IRequestHandler<DeleteBookCommand, List<Book>>
    {
        private readonly IBookRepository _bookRepository;

        public DeleteBookCommandHandler(IBookRepository bookRepository)
        {
            _bookRepository = bookRepository;
        }

        public async Task<List<Book>> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
        {
            var book = await _bookRepository.GetByIdAsync(request.BookId);
            if (book == null)
                throw new KeyNotFoundException("Book not found.");

            await _bookRepository.DeleteAsync(request.BookId);

            var books = await _bookRepository.GetAllAsync();
            return books.ToList();
        }
    }
}
